+++ lucarecord/lib/luca_support/mail.rb
@@ -1,4 +1,6 @@
 require "mail"
+require 'net/smtp'
+require 'openssl'
 require "pathname"
 require "luca_record/io"
 
@@ -8,7 +10,7 @@ module LucaSupport
 
     def initialize(mail=nil, pjdir=nil)
       @pjdir = pjdir || Dir.pwd
-      @config = load_config(@pjdir)
+      @config = self.class.load_config(@pjdir)
       @mail = mail
       set_message_default
       @host = set_host
@@ -16,7 +18,11 @@ module LucaSupport
 
     def deliver
       # mail gem accepts hash for 2nd param, not keywords
-      @mail.delivery_method(:smtp, @host)
+      if conn = conn_with_tls
+        @mail.delivery_method(:smtp_connection, { connection: conn })
+      else
+        @mail.delivery_method(:smtp, @host)
+      end
       @mail.deliver
     end
 
@@ -44,5 +50,28 @@ module LucaSupport
       @mail.from ||= @config.dig("mail", "from")
       @mail.cc ||= @config.dig("mail", "cc")
     end
+
+    private
+
+    def conn_with_tls
+      ca_file = mail_config("ca_file")
+      c_cert_path = mail_config("client_cert")
+      c_key_path = mail_config("client_key")
+      return nil if ca_file.nil? && c_cert_path.nil? && c_key_path.nil?
+
+      tls_ctx = OpenSSL::SSL::SSLContext.new
+      tls_ctx.cert = c_cert_path ?
+                       OpenSSL::X509::Certificate.new(File.read(c_cert_path)) : nil
+      tls_ctx.key = c_key_path ?
+                      OpenSSL::PKey::RSA.new(File.read(c_key_path)) : nil
+      tls_ctx.ca_file = ca_file
+      tls_ctx.verify_mode = OpenSSL::SSL::VERIFY_PEER if ca_file || @host[:openssl_verify_mode]
+
+      conn = Net::SMTP.new(@host[:address], @host[:port])
+      conn.enable_tls(tls_ctx)
+      conn.start(helo: @host[:domain], user: @host[:user_name], password: @host[:password], authtype: @host[:authentication])
+      conn
+    end
+
   end
 end